# docs/openapi.yaml
openapi: 3.1.0
info:
  title: AI Bounties API
  version: 0.6.0
  description: >
    Phase 6 API for BSV AI Bounties. Live document is also served at GET /openapi.json.
    Acceptance kinds: manual, http (JSON APIs), schema, hash, command (alias of hash), llm-judge.
    Auth: real Yours/EC keys BSM-sign challenge `message` (compact base64); demo sha256 only for non-EC keys when AUTH_MODE=demo|both.
servers:
  - url: http://localhost:8787
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: >
        Session token from POST /v1/auth/login.
        Challenge demoHint explains BSM vs demo sha256 for the controller key.
paths:
...
  /v1/bounties:
    get:
      operationId: listBounties
      summary: List bounties
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [open, claimed, submitted, paid, refunded, cancelled]
        - name: category
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 50
      responses:
        '200':
          description: Bounty list
    post:
      operationId: createBounty
      summary: Create / index a bounty (Bearer session required)
      security:
        - bearerAuth: []
      responses:
        '201':
          description: Created
        '401':
          description: Unauthorized — login required
...